refactor for integration with other services
