+++ championship-services/src/main/resources/application.properties
@@ -1,10 +1,8 @@
-server.port=8080
-
-spring.jpa.hibernate.ddl-auto=validate
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/championship_data
+spring.datasource.url=jdbc:mysql://localhost:3306/championship_data
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-#spring.jpa.show-sql: true
+spring.jpa.show-sql: true
+spring.jpa.hibernate.ddl-auto=validate
 
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
